requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosmtplib==3.0.1
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2
qrcode[pil]==7.4.2
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0